test(ripple-effect): migrate
